@@ -4,7 +4,7 @@ import com.healplus.backend.Model.Entity.Patient;
 import com.healplus.backend.Model.Entity.RemoteCheckIn;
 import com.healplus.backend.Service.Monitoring.RemoteMonitoringService;
 import com.healplus.backend.Service.Monitoring.RemoteMonitoringService.RemoteCheckInResponse;
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
