@@ -6,8 +6,8 @@ import com.healplus.backend.security.service.RateLimitService;
 import com.healplus.backend.security.service.SecurityService;
 import com.healplus.backend.security.exception.AccessDeniedException;
 import com.healplus.backend.security.exception.SecurityException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.method.HandlerMethod;
